avance carrito
-validaciones a la hora de crear orden, agregar a carrito,  se supone que todo funciona (ordenes)
-validacion en los actionDelete de todos los controller
-agregacion de alerblock y paginacion
-se puso en español
-se definicio zona horario

+++ controllers/Tr08nhrController.php
@@ -12,6 +12,8 @@ use yii\filters\VerbFilter;
 use yii\filters\AccessControl;
 use yii\web\UploadedFile;
 
+date_default_timezone_set(Yii::$app->params['zonaHorario']);
+
 /**
 * Tr08nhrController implements the CRUD actions for Tr08nhr model.
 */
@@ -164,13 +166,29 @@ class Tr08nhrController extends \yii\web\Controller
   public function actionDelete($id)
   {
     $model = $this->findModel($id);
-    if(file_exists(Yii::getAlias('@app').'/web/uploads/lh/'.$model->ima_08vc)
-    && $model->ima_08vc != null){
-      unlink(Yii::getAlias('@app').'/web/uploads/lh/'.$model->ima_08vc);
-    }
-    $model->delete();
 
-    return $this->redirect(['index']);
+    if($model->tr10hers == null){
+      if(file_exists(Yii::getAlias('@app').'/web/uploads/lh/'.$model->ima_08vc)
+      && $model->ima_08vc != null){
+        unlink(Yii::getAlias('@app').'/web/uploads/lh/'.$model->ima_08vc);
+      }
+      if ($model->delete()) {
+        Yii::$app->getSession()->setFlash('success',
+        '<span class="glyphicon glyphicon-ok-sign"></span> <strong>'.
+        Yii::t('app','Nombre eliminado').'!</strong>');
+        return $this->redirect(['index']);
+      }else{
+        Yii::$app->getSession()->setFlash('error',
+        '<span class="glyphicon glyphicon-bullhorn"></span> <strong>'.
+        Yii::t('app','No se pudo eliminar').'!</strong>');
+        return $this->redirect(['index']);
+      }
+    }else{
+      Yii::$app->getSession()->setFlash('error',
+      '<span class="glyphicon glyphicon-bullhorn"></span> <strong>'.
+      Yii::t('app','No se puede eliminar').'!</strong>');
+      return $this->redirect(['index']);
+    }
   }
 
   /**
